fix searcher-deleter deadlock
Anytime a deleter was waiting it would lock the searcher_mutex to
prevent more searchers from joining, but this also prevented any active
searchers from leaving to make room for the deleter, causing a deadlock.
The fix was to make only searchers lock the searcher_mutex, and for that
we only update searcher_count after we lock the mutex.