fix memory leaks
After building with fsanitize=address found some memory leaks in the
llist_print functio which was not freeing its scratch buffer. As well as
added a llist_free function which frees the whole list. It hasn't been
tested with edge cases like the empty list for example, but it works
fine for how it's used for now.