add separate scheduling logic
Branched run logic into a separate file to allow for more control over
how runs are done. For now the mechanism is pretty simple, you define
the total number of searchers, inserters and deleters and run_cfg starts
one randomly every time, and then waits on all of them. This can be
changed to more sophisticated behavior if need be.