add test directory with simple tests
Used the greatest[1] C testing library because it's a single header
file which can be copied and included and just works.

The first tests are very simple ones which test only the core linked
list functions, but it would be good to add tests which exercise the
worker functions, such as running two llist_inserter_acquires one after
the other and assert that only one gets access to the list, etc.

[1]: https://github.com/silentbicycle/greatest