tests: remove fsanitize=address from tests
While this could be useful, in practice it just introduces noise since
we don't really care about memory leaks, specially because memory leaks
in tests are endemic to the tests (e.g. if a test fails and bails early
it won't free resources), so it's better to just have it disabled.