log: Add time tracking info
Still experimenting with the log structure, thought of logging the time
and thread id of an event, so we can connect wait/enter/leave events
from the same thread, and get stats on how long on average each worker
waited, worked, etc.

Made the llog functions return immediately when LOG is not defined, so
our code can call it normally without worrying about anything. Note that
this means that when LOG is not defined the llog pointer inside llist
will be NULL, which shouldn't be a problem as long as we never access it
directly, only through llog functions from log.h

The -D_POSIX_C_SOURCE=199309L compiler flag is needed so that the
clock_gettime function is correctly defined.